[K8C-41] initial commit for k8ssandra-cluster chart
Until cass-operator 1.5 is released we need to use a custom built image of cass-operator
from master. We are using a copy of the cass-operator chart from the cass-operator master
branch in order to get the changes in https://github.com/datastax/cass-operator/pull/296.
Due to changes in the CRD, we have to use an operator image > 1.4.1.