Fix doc builds [K8SSAND-566] Corrects the failing doc builds by adding in the missing peer dependency of the newer libraries being used.
This also removes the intermediate `install-npm-tools` script in favor of declaration within the package.json file for the docs module.

I tested the workflow in my fork and you can see [here](https://github.com/jdonenine/k8ssandra/runs/2794902349?check_suite_focus=true) that it built successfully.  The deploy fails because it was in my fork without the required secrets.