Add workflow for creating dependency upgrade PRs
This workflow will trigger on a schedule and check the latest versions available in Docker Hub for Stargate, Reaper and Medusa. If the current version in use in values.yaml is different, a PR will be created per dependency and integration tests will run to verify the last version doesn't break CI.
Add GHA script to set tools versions variables.