fix bug in upgrade test
In k8ssandra 1.3.0 we changed the default version of C* to 4.0. The
TestUpgradeScenario test first installs k8ssandra using the default version for
C* which is now 4.0. It then upgrades to 3.11.10. This is causing instability
in the C* cluster.

This commit is intended as a band aide of sorts. The upgrade tests need some
refactoring now that 4.0 is the default version for C* but also to simplify
the code some.