do not change num_tokens when upgrading (#1031)
* do not change num_tokens when upgrading

* fix indentation

* check if num_tokens is set on upgrade

* apply value from previous install when num_tokens is not set

* add some comments

* using the existing num_tokens if possible, add validation check

* fix lint error

* fix typo

* fix bug in upgrade test

In k8ssandra 1.3.0 we changed the default version of C* to 4.0. The
TestUpgradeScenario test first installs k8ssandra using the default version for
C* which is now 4.0. It then upgrades to 3.11.10. This is causing instability
in the C* cluster.

This commit is intended as a band aide of sorts. The upgrade tests need some
refactoring now that 4.0 is the default version for C* but also to simplify
the code some.