Fix upgrade tests (main branch) (#1073)
* make sure a change is applied to the CassandraDatacenter

We have been seeing failures in the upgrade tests on the resource version
check. Now that 4.0 is the default, the second upgrade subtest that upgrades
from latest to latest doesn't actually change the CassandraDatacenter. This
commit adds an explicit change to guarantee an update.

* Need to make sure there is an update to the CassandraDatacenter

This is the same issue as the previous commit but the RestoreAfterUpgrade
test. We use modified chart that guarantees an update.

* only change overrides file for minio which is backend used in TestRestoreAfterUpgrade