Preparing the `docs` branch for production release
This removes the `temp` deployments from the `docs` branch and effectively makes the `docs` branch the offical documentation branch now, not main.

Once this is merged a push of any tags to the `docs` branch with a structure like:

* `docs-v*.*.*`
* `docs-v*.*.*-*`

Will prompt a publish of the content on the `docs` branch to https://docs.k8ssandra.io.

Simply merging this will NOT cause any changes to the main docs.k8ssandra.io site itself.