Remove docs from main branch
The docs content will now live solely in the docs related branches and not within main.

This change itself does not impact https://docs.k8ssandra.io -- that will only be updated once tags are pushed related to other pending PRs for the release.

There will be a follow up PR that adds the /docs directory to gitignore because the directory will still be used for the temporary generation of helm docs.